+++ packages/block-editor/src/private-apis.js
@@ -76,6 +76,7 @@ import useRemoteUrlData from './components/link-control/use-rich-url-data';
 import { PrivateBlockContext } from './components/block-list/private-block-context';
 import useListViewPanelState from './components/use-list-view-panel-state';
 import InnerContent from './components/inner-content';
+import InterceptedLinkPopover from './components/intercepted-link-popover';
 import { useNativeUndo, usesNativeUndo } from './utils/native-undo';
 import {
 	isHashLink,
@@ -158,6 +159,7 @@ lock( privateApis, {
 	isHashLink,
 	isRelativePath,
 	InnerContent,
+	InterceptedLinkPopover,
 	useNativeUndo,
 	usesNativeUndo,
 	isElementVisible,